ref: remove models directory
