--- src/models/IUser.ts
@@ -1,9 +1,0 @@
-export interface IUser {
-	// id: number | null;
-	// auth_token: string | null;
-	first_name: string;
-	second_name: string;
-	phone: string;
-	email: string;
-	photo: string;
-}
